# src/main/java/com/portifolio/fintrack/controller/AuthController.java
package com.portifolio.fintrack.controller;

import com.portifolio.fintrack.dto.AuthRequest;
import com.portifolio.fintrack.dto.AuthResponse;
import com.portifolio.fintrack.dto.MensagemResponse;
import com.portifolio.fintrack.dto.PerfilRequest;
import com.portifolio.fintrack.dto.RecuperacaoSenhaRequest;
import com.portifolio.fintrack.dto.RedefinirSenhaRequest;
import com.portifolio.fintrack.dto.RegistroRequest;
import com.portifolio.fintrack.service.AuthService;
import com.portifolio.fintrack.service.JwtService.JwtUsuario;
import jakarta.validation.Valid;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

@RestController
@RequestMapping("/auth")
@CrossOrigin(origins = "*")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/registrar")
    public AuthResponse registrar(@Valid @RequestBody RegistroRequest request) {
        return authService.registrar(request);
    }

    @PostMapping("/login")
    public AuthResponse login(@Valid @RequestBody AuthRequest request) {
        return authService.login(request);
    }

    @PostMapping("/recuperar-senha")
    public MensagemResponse recuperarSenha(@Valid @RequestBody RecuperacaoSenhaRequest request) {
        return authService.solicitarRecuperacao(request);
    }

    @PostMapping("/redefinir-senha")
    public MensagemResponse redefinirSenha(@Valid @RequestBody RedefinirSenhaRequest request) {
        return authService.redefinirSenha(request);
    }

    @PutMapping("/perfil")
    public AuthResponse atualizarPerfil(@AuthenticationPrincipal JwtUsuario usuario, @Valid @RequestBody PerfilRequest request) {
        return authService.atualizarPerfil(usuario.usuarioId(), request);
    }

    @PostMapping("/avatar")
    public AuthResponse salvarAvatar(@AuthenticationPrincipal JwtUsuario usuario, @RequestParam("avatar") MultipartFile avatar) {
        return authService.salvarAvatar(usuario.usuarioId(), avatar);
    }
}